requirements.txt
# Антивирусная база данных - требования
# Python версия: 3.8+

# 1. Графический интерфейс
PyQt6==6.6.1                    # Основная библиотека GUI
PyQt6-Qt6==6.6.1               # Qt6 бинарные файлы
PyQt6-sip==13.6.0              # Система сигналов PyQt

# 2. Работа с базой данных
mysql-connector-python==8.2.0  # Драйвер MySQL для Python
peewee==3.17.0                  # Простой и легкий ORM

# 3. Создание отчетов
xlsxwriter==3.1.9              # Создание Excel файлов

# 4. Визуализация данных
matplotlib==3.8.2              # Графики и диаграммы

# 5. Вспомогательные библиотеки
python-dateutil==2.8.2         # Работа с датами
